// src/models/type-observable.ts
/**
 * Tipo de Observable
 */
export interface TypeObservable {
  /**
   * Nombre del observable
   */
  name: string;
  /**
   * Titulo que va mostrar
   */
  title: string;
  /**
   * Informacion del observable
   */
  info: string;
  /**
   * Url de la imagen del codigo
   */
  urlImage: string;
}